+++ src/main/java/com/plaid/client/PlaidClient.java
@@ -5,10 +5,12 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.plaid.client.internal.Util;
 import com.plaid.client.internal.gson.CredentialInjectingTypeAdapterFactory;
-import com.plaid.client.internal.gson.ImmutableListTypeAdapterFactory;
-import com.plaid.client.internal.gson.BaseOptionsSerializer;
+import com.plaid.client.internal.gson.ImmutableListStripUnknownEnumsTypeAdapterFactory;
+import com.plaid.client.internal.gson.InvestmentsTransactionsBaseOptionsSerializer;
+import com.plaid.client.internal.gson.TransactionsBaseOptionsSerializer;
 import com.plaid.client.internal.gson.OptionalTypeAdapterFactory;
 import com.plaid.client.internal.gson.RequiredFieldTypeAdapterFactory;
+import com.plaid.client.request.InvestmentsTransactionsGetRequest;
 import com.plaid.client.request.TransactionsGetRequest;
 import com.plaid.client.response.ErrorResponse;
 import java.io.IOException;
@@ -136,17 +138,18 @@ public final class PlaidClient {
     private final OkHttpClient.Builder okHttpClientBuilder;
     private String baseUrl;
     private HttpLoggingInterceptor.Level httpLogLevel;
-    private long readTimeoutSeconds;
-    private long connectTimeoutSeconds;
     private String publicKey;
 
     private String clientId;
     private String secret;
 
     private Builder() {
-      this.okHttpClientBuilder = new OkHttpClient.Builder();
-      this.readTimeoutSeconds = DEFAULT_READ_TIMEOUT_SECONDS;
-      this.connectTimeoutSeconds = DEFAULT_CONNECT_TIMEOUT_SECONDS;
+      this.okHttpClientBuilder = new OkHttpClient.Builder()
+        .readTimeout(DEFAULT_READ_TIMEOUT_SECONDS, TimeUnit.SECONDS)
+        .connectTimeout(DEFAULT_CONNECT_TIMEOUT_SECONDS, TimeUnit.SECONDS)
+        .followSslRedirects(false)
+        .addInterceptor(new PlaidApiHeadersInterceptor())
+        .connectionSpecs(Collections.singletonList(CONNECTION_SPEC));
     }
 
     /**
@@ -181,19 +184,13 @@ public final class PlaidClient {
           new CredentialInjectingTypeAdapterFactory(publicKey, clientId, secret))
         .registerTypeAdapterFactory(new RequiredFieldTypeAdapterFactory())
         .registerTypeAdapterFactory(new OptionalTypeAdapterFactory())
-        .registerTypeAdapterFactory(new ImmutableListTypeAdapterFactory())
-        .registerTypeAdapter(TransactionsGetRequest.BaseOptions.class, new BaseOptionsSerializer())
+        .registerTypeAdapterFactory(new ImmutableListStripUnknownEnumsTypeAdapterFactory())
+        .registerTypeAdapter(TransactionsGetRequest.BaseOptions.class, new TransactionsBaseOptionsSerializer())
+        .registerTypeAdapter(InvestmentsTransactionsGetRequest.BaseOptions.class, new InvestmentsTransactionsBaseOptionsSerializer())
         .create();
     }
 
     private OkHttpClient buildOkHttpClient() {
-      okHttpClientBuilder
-        .readTimeout(readTimeoutSeconds, TimeUnit.SECONDS)
-        .connectTimeout(connectTimeoutSeconds, TimeUnit.SECONDS)
-        .followSslRedirects(false)
-        .addInterceptor(new PlaidApiHeadersInterceptor())
-        .connectionSpecs(Collections.singletonList(CONNECTION_SPEC));
-
       if (httpLogLevel != null) {
         okHttpClientBuilder.addInterceptor(new HttpLoggingInterceptor().setLevel(httpLogLevel));
       }