+++ src/main/java/com/plaid/client/response/IdentityGetResponse.java
@@ -1,41 +1,45 @@
 package com.plaid.client.response;
 
+import java.util.Objects;
 import java.util.List;
+import java.util.Objects;
 
 public final class IdentityGetResponse extends BaseResponse {
-  private List<Account> accounts;
   private ItemStatus item;
-  private Identity identity;
+  private List<AccountWithOwners> accounts;
 
   public ItemStatus getItem() {
     return item;
   }
 
-  public Identity getIdentity() {
-    return identity;
-  }
-
-  public List<Account> getAccounts() {
+  public List<AccountWithOwners> getAccounts() {
     return accounts;
   }
 
+  public static final class AccountWithOwners extends Account {
+    private List<Identity> owners;
+
+    public List<Identity> getOwners() {
+      return owners;
+    }
+  }
+
   public static final class Identity {
-    private List<String> names;
-    private List<Email> emails;
     private List<Address> addresses;
+    private List<Email> emails;
+    private List<String> names;
     private List<PhoneNumber> phoneNumbers;
 
-    public List<String> getNames() {
-      return names;
+    public List<Address> getAddresses() {
+      return addresses;
     }
 
     public List<Email> getEmails() {
       return emails;
     }
 
-
-    public List<Address> getAddresses() {
-      return addresses;
+    public List<String> getNames() {
+      return names;
     }
 
     public List<PhoneNumber> getPhoneNumbers() {
@@ -43,47 +47,43 @@ public final class IdentityGetResponse extends BaseResponse {
     }
   }
 
-  public static final class Email {
-    private Boolean primary;
-    private String data;
-    private String type;
-
-    public Boolean isPrimary() {
-      return primary;
-    }
-
-    public String getData() {
-      return data;
-    }
-
-    public String getType() {
-      return type;
-    }
-  }
-
   public static final class Address {
-    private List<String> accounts;
-    private Boolean primary;
     private AddressData data;
-
-    public Boolean isPrimary() {
-      return primary;
-    }
-
-    public List<String> getAccounts() {
-      return accounts;
-    }
+    private boolean primary;
 
     public AddressData getData() {
       return data;
+    }
+
+    public boolean isPrimary() {
+      return primary;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+      if (this == obj) {
+        return true;
+      }
+      if (obj == null || getClass() != obj.getClass()) {
+        return false;
+      }
+      Address address = (Address) obj;
+      return primary == address.primary &&
+          Objects.equals(data, address.data);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(data, primary);
     }
   }
 
   public static final class AddressData {
     private String street;
     private String city;
-    private String state;
-    private String zip;
+    private String region;
+    private String postalCode;
+    private String country;
 
     public String getStreet() {
       return street;
@@ -93,30 +93,112 @@ public final class IdentityGetResponse extends BaseResponse {
       return city;
     }
 
-    public String getState() {
-      return state;
+    public String getRegion() {
+      return region;
     }
 
-    public String getZip() {
-      return zip;
+    public String getPostalCode() {
+      return postalCode;
+    }
+
+    public String getCountry() {
+      return country;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+      if (this == obj) {
+        return true;
+      }
+      if (obj == null || getClass() != obj.getClass()) {
+        return false;
+      }
+      AddressData addressData = (AddressData) obj;
+      return Objects.equals(street, addressData.street) &&
+          Objects.equals(city, addressData.city) &&
+          Objects.equals(region, addressData.region) &&
+          Objects.equals(postalCode, addressData.postalCode) &&
+          Objects.equals(country, addressData.country);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(street, city, region, postalCode, country);
     }
   }
 
-  public static final class PhoneNumber {
-    private Boolean primary;
+  public static final class Email {
     private String data;
+    private boolean primary;
     private String type;
-
-    public Boolean isPrimary() {
-      return primary;
-    }
 
     public String getData() {
       return data;
     }
 
+    public boolean isPrimary() {
+      return primary;
+    }
+
     public String getType() {
       return type;
     }
+
+    @Override
+    public boolean equals(Object obj) {
+      if (this == obj) {
+        return true;
+      }
+      if (obj == null || getClass() != obj.getClass()) {
+        return false;
+      }
+      Email email = (Email) obj;
+      return primary == email.primary &&
+          Objects.equals(data, email.data) &&
+          Objects.equals(type, email.type);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(data, primary, type);
+    }
   }
+
+  public static final class PhoneNumber {
+    private String data;
+    private boolean primary;
+    private String type;
+
+    public String getData() {
+      return data;
+    }
+
+    public boolean isPrimary() {
+      return primary;
+    }
+
+    public String getType() {
+      return type;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+      if (this == obj) {
+        return true;
+      }
+      if (obj == null || getClass() != obj.getClass()) {
+        return false;
+      }
+      PhoneNumber that = (PhoneNumber) obj;
+      return primary == that.primary &&
+          Objects.equals(data, that.data) &&
+          Objects.equals(type, that.type);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(data, primary, type);
+    }
+  }
+
 }